new Light struct that to replace Strip. unlisted old mods
I unlisted the old mods as at my current rate I am really doing a rewrite and
keeping them in sync with the new code I was writing was more of a hassle then it
was worth.

I did not delete the file though so I can still use them as reference material.
during the rewrite.