Moved the mock boundary from the lights to pins using hardware mod
Instead of mocking the lights (known as strips in older version),I decided to mock the Pins in side of them.
This makes the amount of code that is untestable quite a bit smaller and
contains most of it to the hardware mod rather then spreading it throughout the code.