corrected changes that would not compile on the pi.
There is one crate that will not compile if not of the pi.
I have isolated it useing cfg tages so that bit of the code is avoided on my mac.
So when I need to check for bugs with that code I will have to ssh into the
pi.(like right now)